Use a static action in events
When we filter event, we do an exact match. Thus if the Action content
is dynamic, we are filtering out some events. The Attributes field is
there to pass additional data with events.

Fixing exec_create, exec_start and health_status event in that matter. A
possible follow-up of this would be to force a type for Action in these
methods to make sure we pass predefined actions.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>